Two fixes
First, make NewMarshaler wait until all rationalizers have become
healthy. This does introduce the possibility that we never return,
but that would imply bugs in the system and is strictly better than a
rationalizer that isn't healthy.

Second, there was an edge case where consuming from a partition -- and
falling behind while doing so -- could lead to getting 'offset out of
range' errors when Kafka rolls a log segment. In that case, we now
immediately give up on the partition claim.