Fix rationalizer wait
The "mark when done" was firing when we consumed the latest message
from Kafka, but since we turn around and enqueue the message in our own
channel it means that the engine hadn't actually processed it yet. This
fixes that bug.

This also fixes load/lag calculation in the Consumer which was bogusly
counting even unclaimed partitions.

There is also some cleanup to make more consistent code.