Working on tests
Moved claim tests (lower level things) into a ClaimSuite, and also
figured out how to make sure to not have to use time as a way to wait
for the rationalizer. This increases the stability and speed of the
tests.

Some small tweaks were made to various parts of the system on messages
to convert errors to warnings where we were calling non-critical states
errors.