Add committed offset usage
Given the Marshal coordination topic's retention time might be set
fairly short (and should be, in general), we run the risk of offset
memory disappearing if a consumer is down for longer than the retention
period. This is suboptimal.

This implements usage of Kafka's offset coordinator interface to enable
us to store and retrieve offsets. The implemented behavior will take the
max of any offset retrieved from the coordination topic or Kafka.

Fixes #6.