Add real at-least-once
This refactors the way we handle consumption to be channel based. The
client requests a channel from the Consumer, and all claim objects will
write to that channel.

Messages that go into the channel were obtained in a valid manner (i.e.,
but an unexpired claim), but we do not consider them to be "finished"
until the client calls Commit and passes the message back to us.

Once Commit is called, we update the internal state and allow the
heartbeat to move the current offset forward.

Fixes #18.