Fix getTopicState
This method is used by everybody to get the state object to update
within the rationalizer. The problem was that it didn't consider the
groupID and always used Marshal's current GroupID. This meant that, for
all running Marshal clients, they shared the same GroupID in practice.

This is obviously very bad.

This fixes the method to actually properly use GroupID to maintain state
and adds a test to ensure that claims by other users don't affect our
own state.