Commit Upon Terminating Consumer
We need to update the currrent offset during both heartbeats and
when we terminate. That way we can commit the offsets back to Kafka.
This changes does that. It also includes a breaking change to the API
by changing Terminate to have an extra flag that indicates whether we
should release the partition or just commit the offsets without releasing
the partition. It should be noted that we will need to update some
dependent code once this is checked-in internally.

Added tests for the new code and made sure things pass.