Limit the rationalizer history view
This makes the rationalizers only go back 100000 offsets. This number is
derived by a best guess at a high level of load for a Marshal partition.
I.e., this number is based on:

* 100,000 / 120 seconds = ~833 QPS to the partition
* In standard load, all claims generate 1 Heartbeat per ~30 seconds
* That means 24,990 partitions need to be claimed _per Marshal partition_
  for this threshold to be exceeded
* The Marshal topic can be partitioned heavily (we use 50 at Dropbox)
* Kafka itself only handles ~10,000 partitions per cluster, so exceeding
  this threshold would be pretty spectacular

Given all this, I believe 100,000 offsets to be safe and sane with
plenty of headroom.

The worst case behavior if this number is insufficient is that we will
claim a partition someone else owns. In that case, the other consumer
will detect that someone conflicted and it will stop consuming. This
will lead to at worse a little more double-consumption and partition
churn, which should be caught with monitoring.