Improve locking
We were over-broad in our locking. This puts the locks around the
accesses to the data structure so we protect them, but then releases
them before we talk to Kafka. This improves heartbeat throughput.

We also remove all locking from CommitOffsets since it just sends a
message to Kafka and doesn't use any local data.