Remove log.Fatalf in double-claim
This fixes the plumbing such that when Marshal is terminated, it is
able to pass that termination on to the consumers. This also fixes some
further plumbing to ensure that we close the channels when we terminate.

This makes Marshal less likely to blow up your binary when it gets mad,
which is generally considered a good thing.