Create 'cluster' abstraction
A long-standing issue with Marshal has been that every time you call
NewMarshaler you end up with a new kafka.Broker which spins up an
entirely new set of rationalizers. This can be quite the large amount of
work for no real benefit. The system was conceived with the idea that
it could be independent of who was using it, so this diff finally does
that.

The abstraction now is:

* Marshaler -> created once per (client, group) and is in charge of
  handling the state for that level. This is the main thing you create
  and interact with as a user.
* cluster -> behind the scenes, one of these is created per unique
  set of (broker, broker, broker...) ipports. This means that as long
  as you are consistent with how you pass brokers to NewMarshaler,
  they will share a cluster under the hood for performance.
* Consumer -> actually instantiated per (Marshaler, topics) that you
  want to consume data from.

This should allow heavy users of Marshal to have some much greater
efficiencies.