Adjust health check and add verbosity
This updates the health check by predicting a consumer offset at the
next heartbeat period. This is done by taking the current velocity and
adding it to the current offset. If this is greater than the end of the
partition, we consider the consumer healthy. This should address the
issues with consumer health oscillation in some cases.

Additionally, this adds a number of metrics to the heartbeat logs and in
a few more case statements to hopefully help people debug issues more
easily.