Add CommitToken and CommitByToken
This switches us to using our own Message type instead of the Kafka
library which allows us to extend it to support a CommitToken method.
This is used to convert the message into a small token which can be used
to reference that message later so you don't have to keep the message
around. You can then pass this to a consumer's CommitByToken method.

This flow is designed for memory utilization and should be used instead
of the normal Commit flow which takes messages.