marshal sending partial topic updates periodically
Issue was a regression with 6f82eaf where we were sending partial
topic updates everytime were looping through claimTopics. This was
happening because we were starting from an empty list and building
and sending notifications accordingly. The fix is to use previous list
and update it and send notifications as necessary. I've updated
the test to cover this case too and have tested it on a machine
in prod this time to verify