Release topic in topic mode
Sometimes we need to release a single partition in topic claim mode for
one reason or another (failed heartbeat, somebody else claimed, etc). It
was previously possible to end up in a split brain situation where we
release partitions >0 and nobody else will ever claim them due to the
way topic claim works.

This update makes it so that if we're in topic claim mode and we release
any partition, we will kick off releases for all of them.