Add options to cluster dialer
This allows passing in options to the cluster dialer used for advanced
use cases. This can be used to configure under the hood how many of
various resources Marshal uses as well as certain timeouts.

These can be used if you need to change behaviors, although we recommend
carefully testing any changes you make so as to avoid any adverse
affects.