Add new ReleaseGroup message
Fix message return type

Add MsgExpireTime field to ReleaseGroup, remove AdminClientID field

Rename LastOffset -> CurrentOffset

The former name was a misnomer. This offset is the 'current offset' that
we are using to refer to the offset we are next going to be consuming.

Revert to UNIX timestamps

Add new ReleaseGroup message

Fix message return type

Add MsgExpireTime field to ReleaseGroup, remove AdminClientID field

Rename LastOffset -> CurrentOffset

The former name was a misnomer. This offset is the 'current offset' that
we are using to refer to the offset we are next going to be consuming.

Revert to UNIX timestamps