Fix refactor problem
To make tests more reliable I made health check loop not start
automatically when the claim starts. It's now the reponsibility of the
consumer to start the health check loop. I forgot to make fast reclaims
start the loop, leading to an issue where 2 minutes after startup we
would see that the claim has gone away and would attempt to claim it
again, thereby double-claiming and failing.

This fixes the 2 minute crash but it still shouldn't be double-claiming
something it already "knows" it has. Next diff.