Remove internal 'claimed' marker
The claim type maintained an internal "am I claimed or not" marker which
it used to do some optimizations related to releasing so it didn't have
to talk to the Marshal object all of the time. This maintaining the
"same" state in two places leads to edge cases and can, in cases, cause
the fatal "internal double-claim" issue when an existing claim doesn't
realize it has died.

This diff removes the internal representation of claim state and uses
the global data. This means that calling release on a claim does not
immediately reflect the release until the rationalizers have received
the update and processed it.

This diff also fixes an issue where newClaim's returned value wasn't
being checked. The function can return nil.