Fix issue with pv=0 health checks
In the case where producers stop writing to a partition, the "partition
velocity" would drop to 0. This basically enabled consumers to always be
healthy -- even if they would stop consuming while behind. Instead, we
would expect a stalled consumer to become unhealthy if they can't catch
up -- even if the producers aren't writing.