Fixes a few issues
First, primarily, this fixes the internal double claim which was firing
in a racy manner. When we would be blocked on obtaining a lock long
enough that we saw the ClaimPartition message arrive, we could "detect"
an internal double-claim erroneously. This is now fixed via:

The claim object used to have this concept of 'Claimed' and 'Terminated'
which meant different things but it was subtle. The double-claim was
attempting to determine if we had two claim objects referencing the
same data but using the 'Claimed' state (which uses the rationalizer
global world state) is inherently racy since the world may change while
we're waiting on a lock. We've removed this notion of 'Claimed' from the
'claim' object (follow me?) because I assert:

* Asking a claim object if the thing it refers to is globally Claimed
doesn't make much sense because we are usually saying 'I would like
to perform some operation on you, may I?' and that's more properly a
reference to Terminated,

* A claim will self-Terminate during its normal heartbeat flow long
before Claimed will return false so it's a quicker indicator,

Therefore I believe that using Terminated instead of a mix is strictly
better from a performance perspective and leaves us open to fewer bugs,
plus it makes it very clear what you're asking a claim object about and
when you're asking the Marshaler for world-state information.

This diff also includes some cleanup (I got annoyed at the 'isFoo'
naming convention that was used in one place).