Change offset commit order
In the old day when offsets only went forward it was fine to release and
then commit. In the new world where offsets might go backwards, however,
this exposes us to a problem where we release and someone else claims
the partition before we've committed the new (earlier) offset. Since
the consumer takes the MAX of (committed, marshal) offsets, this leads
consumers to wrongly go forward during a rewind operation.

This fixes the ordering which will address that condition. This also
makes the CommitOffset method stop suppressing errors and updates the
callers to suppress errors (if they want to).