Use offsets in priority order instead of MAX
The old logic just took the largest offset no matter what, but this
causes issues with rewinding since we can end up committing an offset
but failing to release, for example. This makes the system prefer to
always use the offset in the Marshal topic and only if we don't have one
will it fall back to using the committed offset.